+++ pkg/api/resources/friends/get_followings_test.go
@@ -56,7 +56,7 @@ func TestGetFollowings(t *testing.T) {
 			WithSortOrderDesc()
 
 		params := builder.Build()
-		assert.Equal(t, uint64(utils.SampleUserID1), params.UserID)
+		assert.Equal(t, utils.SampleUserID1, params.UserID)
 		assert.Equal(t, uint64(50), params.Limit)
 		assert.Equal(t, "someCursor", params.Cursor)
 		assert.Equal(t, types.SortOrderDesc, params.SortOrder)
