+++ pkg/api/resources/users/username_history_test.go
@@ -58,7 +58,7 @@ func TestGetUsernameHistory(t *testing.T) {
 			WithSortOrderDesc()
 
 		params := builder.Build()
-		assert.Equal(t, uint64(utils.SampleUserID1), params.UserID)
+		assert.Equal(t, utils.SampleUserID1, params.UserID)
 		assert.Equal(t, uint64(50), params.Limit)
 		assert.Equal(t, "someCursor", params.Cursor)
 		assert.Equal(t, types.SortOrderDesc, params.SortOrder)
